Initial cordova/ionic checkin
This time, we will avoid the challenges of the prior time, and check in the
migration in stages. This checks in the initial blank implementation, along
with the SQLite plugin.

We will check in the android and ios native code in separate checkins because
its going to be a bunch of fiddling to get them to work correctly, in
particular, because it is unsure how much native code to check in.