Woot! It werks!!
Hooked up gnarly angular stuff to get everything to work.
Angular isn't so bad if you understand it.
The main thing to understand is that javascript is all about callbacks and asynchronous execution.
Once you understand that, you just need to work on looking up how to call
various refresh/synchronization mechanisms.

At this point, I can display the 27 trips in the sample prepopulated database.
Unfortunately, none of them have a predicted mode, so there are no confidence
levels, and all of them are walking trips.

But I claim that this is now working, and plan to integrate it into the real
app soon.